app/requirements.txt
celery
deepspeech-tflite==0.8.2
Flask
opencv-python
